+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- Auto-defined assertions based on user-defined computed getter properties
+
 ### Fixed
 
 - Assigning to getters using the legacy decorator